Create Disemvowel_Trolls.
Originally tried a range based for loop and replacing the vowel with a blank line. However, I noticed that the string passed in was const. I looked up ways to erase certain characters in a string and while there are built in functions that do this, I wanted to do it without using these functions to better understand how to do it.